app/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.4.0
audioread==3.0.1
Authlib==1.6.5
bitsandbytes==0.46.0
braceexpand==0.1.7
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.3
cryptography==46.0.2
cuda-bindings==13.0.2
cuda-pathfinder==1.3.1
cycler==0.12.1
cytoolz==1.0.1
datasets==4.2.0
decorator==5.2.1
dill==0.4.0
docopt==0.6.2
dparse==0.6.4
editdistance==0.8.1
einops==0.8.1
executing==2.2.1
fastapi==0.119.0
fiddle==0.3.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.0
indic_numtowords==1.1.0
inflect==7.5.0
intervaltree==3.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.2
julius==0.2.7
kaldi-python-io==1.2.2
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.27.1
lhotse==1.31.1
libcst==1.8.5
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.2
lilcom==1.8.1
llvmlite==0.45.1
loguru==0.7.3
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.0.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.1.6
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
nemo-toolkit==2.5.0
networkx==3.5
nltk==3.9.2
num2words==0.5.14
numba==0.62.1
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnx==1.19.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
optuna==4.5.0
packaging==24.2
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
plac==1.4.5
platformdirs==4.5.0
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.5
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote-audio==4.0.1
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pyarrow==21.0.0
pybind11==3.0.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pydub==0.25.1
Pygments==2.19.2
pyloudnorm==0.1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
resampy==0.4.3
rich==14.2.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
sacremoses==0.1.1
safetensors==0.6.2
safety==3.6.2
safety-schemas==0.0.16
scikit-learn==1.7.2
scipy==1.16.2
sentencepiece==0.2.1
sentry-sdk==2.41.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
tokenizers==0.21.4
tomlkit==0.13.3
toolz==1.0.0
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchcodec==0.7.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.3
triton==3.4.0
typeguard==4.4.4
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
wandb==0.22.2
watchfiles==1.1.0
wcwidth==0.2.14
webdataset==1.0.2
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
whisper_normalizer==0.1.12
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
